Put FakeRepository behind test cfg
